Move TestTensorInstanceMethods to tests.tensor.test_var
