+++ tests/tensor/test_basic.py
@@ -6,7 +6,6 @@ from tempfile import mkstemp
 
 import numpy as np
 import pytest
-from numpy.testing import assert_array_equal
 
 import aesara
 import aesara.scalar as aes
@@ -115,7 +114,6 @@ from aesara.tensor.type import (
     ivector,
     lscalar,
     lvector,
-    matrices,
     matrix,
     row,
     scalar,
@@ -3958,75 +3956,6 @@ class TestInferShape(utt.InferShapeTester):
             [0, 0, 1],
             ARange,
         )
-
-
-class TestTensorInstanceMethods:
-    def setup_method(self):
-        self.vars = matrices("X", "Y")
-        self.vals = [m.astype(config.floatX) for m in [random(2, 2), random(2, 2)]]
-
-    def test_repeat(self):
-        X, _ = self.vars
-        x, _ = self.vals
-        assert_array_equal(X.repeat(2).eval({X: x}), x.repeat(2))
-
-    def test_trace(self):
-        X, _ = self.vars
-        x, _ = self.vals
-        assert_array_equal(X.trace().eval({X: x}), x.trace())
-
-    def test_ravel(self):
-        X, _ = self.vars
-        x, _ = self.vals
-        assert_array_equal(X.ravel().eval({X: x}), x.ravel())
-
-    def test_diagonal(self):
-        X, _ = self.vars
-        x, _ = self.vals
-        assert_array_equal(X.diagonal().eval({X: x}), x.diagonal())
-        assert_array_equal(X.diagonal(1).eval({X: x}), x.diagonal(1))
-        assert_array_equal(X.diagonal(-1).eval({X: x}), x.diagonal(-1))
-        for offset, axis1, axis2 in [(1, 0, 1), (-1, 0, 1), (0, 1, 0), (-2, 1, 0)]:
-            assert_array_equal(
-                X.diagonal(offset, axis1, axis2).eval({X: x}),
-                x.diagonal(offset, axis1, axis2),
-            )
-
-    def test_take(self):
-        X, _ = self.vars
-        x, _ = self.vals
-        indices = [1, 0, 3]
-        assert_array_equal(X.take(indices).eval({X: x}), x.take(indices))
-        indices = [1, 0, 1]
-        assert_array_equal(X.take(indices, 1).eval({X: x}), x.take(indices, 1))
-        indices = np.array([-10, 5, 12], dtype="int32")
-        assert_array_equal(
-            X.take(indices, 1, mode="wrap").eval({X: x}),
-            x.take(indices, 1, mode="wrap"),
-        )
-        assert_array_equal(
-            X.take(indices, -1, mode="wrap").eval({X: x}),
-            x.take(indices, -1, mode="wrap"),
-        )
-        assert_array_equal(
-            X.take(indices, 1, mode="clip").eval({X: x}),
-            x.take(indices, 1, mode="clip"),
-        )
-        assert_array_equal(
-            X.take(indices, -1, mode="clip").eval({X: x}),
-            x.take(indices, -1, mode="clip"),
-        )
-        # Test error handling
-        with pytest.raises(IndexError):
-            X.take(indices).eval({X: x})
-        with pytest.raises(IndexError):
-            (2 * X.take(indices)).eval({X: x})
-        with pytest.raises(TypeError):
-            X.take([0.0])
-        indices = [[1, 0, 1], [0, 1, 1]]
-        assert_array_equal(X.take(indices, 1).eval({X: x}), x.take(indices, 1))
-        # Test equivalent advanced indexing
-        assert_array_equal(X[:, indices].eval({X: x}), x[:, indices])
 
 
 class TestSwapaxes:
